created BodypartFilter reusable component, created helpers.js in utils with capitalizeWords, added Bodypartfilter functionality in CreatePlan.jsx
